requirements.txt
streamlit>=1.32.0
tensorflow-cpu>=2.15.0
numpy>=1.24.0
pillow>=10.0.0
plotly>=5.18.0
pandas>=2.0.0
gdown>=5.1.0
